Finalise transform
